requirements.txt
langchain
streamlit
python-dotenv
PyPDF2
pandas
fpdf
streamlit
google-generativeai





-e .